move conversion logic to a separate class
