+++ spring-cloud-gcp-data-datastore/src/test/java/org/springframework/cloud/gcp/data/datastore/core/convert/EntityPropertyValueProviderTests.java
@@ -28,7 +28,6 @@ import org.junit.Test;
 import org.springframework.cloud.gcp.data.datastore.core.mapping.DatastoreDataException;
 import org.springframework.cloud.gcp.data.datastore.core.mapping.DatastoreMappingContext;
 import org.springframework.cloud.gcp.data.datastore.core.mapping.DatastorePersistentEntity;
-import org.springframework.core.convert.support.DefaultConversionService;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
@@ -38,13 +37,12 @@ public class EntityPropertyValueProviderTests {
 
 	private Datastore datastore;
 
-	private DefaultDatastoreEntityConverter datastoreEntityConverter =
-			new DefaultDatastoreEntityConverter(new DatastoreMappingContext());
+	private TwoStepsConversions twoStepsConversion = new TwoStepsConversions(new DatastoreCustomConversions());
 
 
 	private DatastorePersistentEntity<TestDatastoreItem> persistentEntity =
-			(DatastorePersistentEntity<TestDatastoreItem>) new DatastoreMappingContext()
-			.getPersistentEntity(TestDatastoreItem.class);
+			(DatastorePersistentEntity<TestDatastoreItem>)
+					new DatastoreMappingContext().getPersistentEntity(TestDatastoreItem.class);
 
 	@Before
 	public void setUp() {
@@ -64,9 +62,7 @@ public class EntityPropertyValueProviderTests {
 				.set("blobField", Blob.copyFrom(bytes))
 				.build();
 
-		EntityPropertyValueProvider provider = new EntityPropertyValueProvider(
-				entity, new DefaultConversionService(), new DefaultConversionService(),
-				this.datastoreEntityConverter::getTwoStepsConversion);
+		EntityPropertyValueProvider provider = new EntityPropertyValueProvider(entity, this.twoStepsConversion);
 
 		assertThat((String) provider.getPropertyValue(this.persistentEntity.getPersistentProperty("stringField")))
 				.as("validate string field").isEqualTo("string value");
@@ -91,9 +87,7 @@ public class EntityPropertyValueProviderTests {
 				.set("boolField", 123L)
 				.build();
 
-		EntityPropertyValueProvider provider = new EntityPropertyValueProvider(
-				entity, new DefaultConversionService(), new DefaultConversionService(),
-				this.datastoreEntityConverter::getTwoStepsConversion);
+		EntityPropertyValueProvider provider = new EntityPropertyValueProvider(entity, this.twoStepsConversion);
 
 		provider.getPropertyValue(this.persistentEntity.getPersistentProperty("boolField"));
 	}
